requirements.txt
asgiref==3.2.10
absl-py==0.8.1
astor==0.8.0
backports.weakref==1.0.post1
absl-py==0.10.0
asgiref==3.2.10
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dicom2nifti==2.2.10
Django==3.0.8
django-betterforms==1.2
django-model-utils==4.0.0
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.1
med2image==2.2.4
nibabel==3.1.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pfmisc==1.4.2
Pillow==7.2.0
protobuf==3.13.0
pudb==2019.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.0.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-resize-image==1.1.19
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
SimpleITK==1.2.4
six==1.15.0
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.25.10
urwid==2.1.1
Werkzeug==1.0.1
wrapt==1.12.1
